+++ tests/Method/Sticker/SetStickerPositionInSetTest.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Tests\Method\Sticker;
+
+use PHPUnit\Framework\TestCase;
+use Vjik\TelegramBot\Api\Method\Sticker\SetStickerPositionInSet;
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+
+final class SetStickerPositionInSetTest extends TestCase
+{
+    public function testBase(): void
+    {
+        $method = new SetStickerPositionInSet('sid', 2);
+
+        $this->assertSame(HttpMethod::POST, $method->getHttpMethod());
+        $this->assertSame('setStickerPositionInSet', $method->getApiMethod());
+        $this->assertSame(
+            [
+                'sticker' => 'sid',
+                'position' => 2,
+            ],
+            $method->getData(),
+        );
+    }
+
+    public function testPrepareResult(): void
+    {
+        $method = new SetStickerPositionInSet('sid', 1);
+
+        $preparedResult = $method->prepareResult(true);
+
+        $this->assertTrue($preparedResult);
+    }
+}