+++ CHANGELOG.md
@@ -40,6 +40,8 @@
 - New #54: Add `getCustomEmojiStickers` method.
 - New #55: Add `uploadStickerFile` method.
 - New #57: Add `addStickerToSet`, `deleteStickerFromSet` and `replaceStickerInSet` methods.
+- New #58: Add `setStickerPositionInSet`, `setStickerEmojiList`, `setStickerKeywords` and `setStickerMaskPosition`
+  methods.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.